format

+++ src/Crank.sol
@@ -19,7 +19,7 @@ import "src/ERC20Container.sol";
 contract Crank is IUniswapV3SwapCallback, Ownable, ERC20Container, UUPSUpgradeable {
     struct SwapCallbackData {
         PoolAddress.PoolKey poolKey;
-	uint limit;
+        uint256 limit;
     }
 
     error Limit(); // frontrunning protection
@@ -30,11 +30,12 @@ contract Crank is IUniswapV3SwapCallback, Ownable, ERC20Container, UUPSUpgradeab
     function initialize(address owner, PoolAddressesProvider aave, address univ3Factory) external {
         // will revert if already initialized
         _initializeOwner(owner);
-	_factory = univ3Factory;
-	_aave = aave;
+        _factory = univ3Factory;
+        _aave = aave;
     }
 
     function _authorizeUpgrade(address) internal override onlyOwner {}
+
     uint16 public constant referralCode = 0;
 
     // transient storage callback lock to prevent unauthorized from flash swapping
@@ -43,7 +44,7 @@ contract Crank is IUniswapV3SwapCallback, Ownable, ERC20Container, UUPSUpgradeab
     ERC20 public constant weth = ERC20(0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2);
     ERC20 public constant wsteth = ERC20(0x7f39C581F595B53c5cb19bD0b3f8dA6c935E2Ca0);
 
-    function _swap(bool zeroForOne, int256 amount, uint24 fee, uint160 sqrtPriceLimitX96, uint limit) internal {
+    function _swap(bool zeroForOne, int256 amount, uint24 fee, uint160 sqrtPriceLimitX96, uint256 limit) internal {
         assembly {
             tstore(CALLBACK_LOCK, 1)
         }
@@ -51,22 +52,26 @@ contract Crank is IUniswapV3SwapCallback, Ownable, ERC20Container, UUPSUpgradeab
             PoolAddress.PoolKey({token0: address(wsteth), token1: address(weth), fee: fee});
         IUniswapV3Pool pool = IUniswapV3Pool(PoolAddress.computeAddress(_factory, poolKey));
         pool.swap(
-            address(this), zeroForOne, amount, sqrtPriceLimitX96, abi.encode(SwapCallbackData({poolKey: poolKey, limit: limit}))
+            address(this),
+            zeroForOne,
+            amount,
+            sqrtPriceLimitX96,
+            abi.encode(SwapCallbackData({poolKey: poolKey, limit: limit}))
         );
         assembly {
             tstore(CALLBACK_LOCK, 0)
         }
     }
 
-    function wind(uint256 wstethAmount, uint24 fee, uint160 sqrtPriceLimitX96, uint limit) external onlyOwner {
+    function wind(uint256 wstethAmount, uint24 fee, uint160 sqrtPriceLimitX96, uint256 limit) external onlyOwner {
         _swap(false, -int256(wstethAmount), fee, sqrtPriceLimitX96, limit);
     }
 
-    function unwind(uint256 wstethAmount, uint24 fee, uint160 sqrtPriceLimitX96, uint limit) external onlyOwner {
+    function unwind(uint256 wstethAmount, uint24 fee, uint160 sqrtPriceLimitX96, uint256 limit) external onlyOwner {
         _swap(true, int256(wstethAmount), fee, sqrtPriceLimitX96, limit);
     }
 
-    function close(uint24 fee, uint160 sqrtPriceLimitX96, uint limit) external onlyOwner {
+    function close(uint24 fee, uint160 sqrtPriceLimitX96, uint256 limit) external onlyOwner {
         (,, uint256 wethAmount,,,,,,) = _aave.getPoolDataProvider().getUserReserveData(address(weth), address(this));
         _swap(true, -int256(wethAmount), fee, sqrtPriceLimitX96, limit);
     }
@@ -80,11 +85,11 @@ contract Crank is IUniswapV3SwapCallback, Ownable, ERC20Container, UUPSUpgradeab
         SwapCallbackData memory decoded = abi.decode(data, (SwapCallbackData));
         IUniswapV3Pool pool = IUniswapV3Pool(PoolAddress.computeAddress(_factory, decoded.poolKey));
         require(msg.sender == address(pool));
-	Pool lendingPool = _aave.getPool();
+        Pool lendingPool = _aave.getPool();
 
         if (amount1Delta > 0) {
             uint256 wethAmount = uint256(amount1Delta);
-	    require(wethAmount >= decoded.limit, Limit());
+            require(wethAmount >= decoded.limit, Limit());
             uint256 wstethBalance = wsteth.balanceOf(address(this));
             wsteth.approve(address(lendingPool), wstethBalance);
             // keep 1 wei for the storage slots
@@ -94,7 +99,7 @@ contract Crank is IUniswapV3SwapCallback, Ownable, ERC20Container, UUPSUpgradeab
             weth.transfer(msg.sender, wethAmount);
         } else {
             uint256 wethAmount = uint256(-amount1Delta);
-	    require(wethAmount >= decoded.limit, Limit());
+            require(wethAmount >= decoded.limit, Limit());
             uint256 wstethAmount = uint256(amount0Delta);
             uint256 wethBalance = weth.balanceOf(address(this));
             weth.approve(address(lendingPool), wethBalance);
